refactor: Removed mainSearch, searchStats, and more
This is fantastic.
