+++ src/js/actions/mainSearch.js
@@ -7,17 +7,13 @@ import {
   issueBoomSearch,
   killBoomSearches
 } from "./boomSearches"
-import {clearLogs, receiveLogTuples} from "./logs"
+import {clearLogs} from "./logs"
 import {getInnerTimeWindow, getOuterTimeWindow} from "../reducers/timeWindow"
-import {getMainSearchRequest} from "../reducers/mainSearch"
 import {getSearchProgram} from "../selectors/searchBar"
-import {getStarredLogs} from "../reducers/starredLogs"
 import {pushSearchHistory} from "./searchHistory"
 import {updateTab} from "../actions/view"
 import {validateProgram} from "./searchBar"
-import ParallelSearch from "../models/ParallelSearch"
 import SearchFactory from "../models/searches/SearchFactory"
-import serially from "../lib/serially"
 
 type Options = {
   saveToHistory: boolean
@@ -30,7 +26,6 @@ export const fetchMainSearch = ({
   if (!dispatch(validateProgram())) return
   dispatch(updateTab(state))
   if (saveToHistory) dispatch(pushSearchHistory())
-  if (starredSearch(state)) return showStarred(state, dispatch)
 
   dispatch(killBoomSearches())
   setTimeout(() => {
@@ -45,55 +40,3 @@ export const fetchMainSearch = ({
     searches.forEach(search => dispatch(issueBoomSearch(search)))
   })
 }
-
-const starredSearch = state => {
-  return getSearchProgram(state) === ":starred"
-}
-
-export const killMainSearch = (): Thunk => (dispatch, getState) => {
-  const request = getMainSearchRequest(getState())
-  request && request.kill()
-}
-
-const showStarred = serially(
-  (state, dispatch) => {
-    const starredLogs = getStarredLogs(state)
-    dispatch(requestMainSearch(new ParallelSearch(dispatch, [])))
-    return setTimeout(() => {
-      dispatch(receiveLogTuples([...starredLogs]))
-      dispatch(completeMainSearch())
-    })
-  },
-  id => clearTimeout(id)
-)
-
-export function requestMainSearch(request: ParallelSearch) {
-  return {
-    type: "MAIN_SEARCH_REQUEST",
-    request
-  }
-}
-
-export function spliceMainSearchEvents(index: number) {
-  return {
-    type: "LOGS_SPLICE",
-    index
-  }
-}
-
-export function completeMainSearch() {
-  return {
-    type: "MAIN_SEARCH_COMPLETE"
-  }
-}
-
-export function appendMainSearchQueryProgram(fragment: string) {
-  return {
-    type: "MAIN_SEARCH_QUERY_PROGRAM_APPEND",
-    fragment
-  }
-}
-
-export const clearMainSearch = () => ({
-  type: "MAIN_SEARCH_CLEAR"
-})
