+++ src/js/connectors/XAnalysisViewer.js
@@ -1,11 +1,16 @@
-import {connect} from "react-redux"
-import AnalysisViewer from "../components/AnalysisViewer"
-import {getAnalysis} from "../reducers/analysis"
-import {getMainSearchIsFetching} from "../reducers/mainSearch"
+/* @flow */
 
-const stateToProps = state => ({
+import {connect} from "react-redux"
+
+import type {State} from "../reducers/types"
+import {getAnalysis} from "../reducers/analysis"
+import {getMainSearchIsFetching} from "../selectors/boomSearches"
+import AnalysisViewer from "../components/AnalysisViewer"
+
+const stateToProps = (state: State) => ({
   analysis: getAnalysis(state),
   isFetching: getMainSearchIsFetching(state)
 })
 
+// $FlowFixMe
 export default connect(stateToProps)(AnalysisViewer)
