--- src/js/lib/serially.js
@@ -1,10 +1,0 @@
-export default (run, abort) => {
-  let prev
-  return (...args) => {
-    if (prev) abort(prev)
-
-    const current = run(...args)
-    prev = current
-    return current
-  }
-}
